add default image for new user

+++ src/utils/utils.ts
@@ -5,6 +5,9 @@ import { unlink } from 'fs';
 
 export const clearImage = (imagePath: string, dirname: string) => {
   imagePath = join(dirname, imagePath);
+  if (imagePath.includes("transparent-avatar.png")) {
+    return
+  }
   unlink(imagePath, (err) => {
     if (err)
       throw err
